pyproject.toml
[project]
name = "zettlecast"
version = "0.1.0"
description = "Digital Zettelkasten Middleware - Local-First AI Knowledge System"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Zettlecast Team"}
]

dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    
    # Database
    "lancedb>=0.17.0",
    "pyarrow>=15.0.0,<18.0.0",  # <18 required for NeMo/datasets compatibility
    
    # Embeddings & ML
    "sentence-transformers>=3.3.0",
    "torch>=2.5.0",
    
    # PDF Parsing (Tiered)
    "pypdf>=5.1.0",
    "marker-pdf>=0.3.0",
    
    # Web Scraping
    "trafilatura>=2.0.0",
    "feedparser>=6.0.0",
    
    # Audio Transcription
    "faster-whisper>=1.1.0",
    
    # UI
    "streamlit>=1.40.0",
    "st-link-analysis>=0.1.0",
    
    # Utilities
    "python-multipart>=0.0.17",
    "pyyaml>=6.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

# Mac-specific (Apple Silicon with MLX)
mac = [
    "parakeet-mlx>=0.1.0",
    "mlx-whisper>=0.4.0",
    "pyannote.audio>=3.1.0",
    "torchaudio>=2.0.0",
]

# Windows-specific (NeMo in container)
windows = [
    "docker>=7.0.0",
]

# Full podcast processing (platform-agnostic)
podcast = [
    # Diarization (works on all platforms)
    "pyannote.audio>=3.1.0",
    "speechbrain>=1.0.0",
    # VAD / Audio
    "torchaudio>=2.0.0",
    "pydub>=0.25.0",
]

# NeMo (for direct installation, not recommended - use container instead)
nemo = [
    "nemo_toolkit[asr]>=2.0.0",
    "omegaconf>=2.3.0",
    "whisperx>=3.1.0",
]


[project.scripts]
zettlecast = "zettlecast.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zettlecast"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
